Moved pagination class
- We have moved the pagination class to a new namespace
