+++ app/Http/Controllers/ItemController.php
@@ -16,7 +16,7 @@ use App\Models\Category;
 use App\Models\Item;
 use App\Models\Subcategory;
 use App\Response\Header\Headers;
-use App\Utilities\Pagination as UtilityPagination;
+use App\Response\Pagination as UtilityPagination;
 use Exception;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\JsonResponse;
