+++ app/Http/Controllers/ItemPartialTransferController.php
@@ -12,7 +12,7 @@ use App\Response\Cache;
 use App\Response\Header\Headers;
 use App\Request\Parameter;
 use App\Request\Route;
-use App\Utilities\Pagination as UtilityPagination;
+use App\Response\Pagination as UtilityPagination;
 use App\Request\Validate\ItemPartialTransfer as ItemPartialTransferValidator;
 use Exception;
 use Illuminate\Database\QueryException;
